testPhonemeComparedToLaterIsNegative

+++ src/test/java/org/apache/commons/codec/language/bm/RuleTest.java
@@ -19,7 +19,6 @@ package org.apache.commons.codec.language.bm;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-
 import org.junit.jupiter.api.Test;
 
 /**
@@ -49,8 +48,8 @@ public class RuleTest {
             for (int i = 0; i < phs.length; i++) {
                 for (int j = i + 1; j < phs.length; j++) {
                     final int c = Rule.Phoneme.COMPARATOR.compare(phs[i], phs[j]);
-
-                    assertTrue(Integer.valueOf(c).intValue() < 0,
+    
+                    assertTrue(c < 0,
                             "Comparing " + phs[i].getPhonemeText() + " to " + phs[j].getPhonemeText() + " should be negative");
                 }
             }
